# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Pydantic model for an RFC 7807 problem-details payload.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation of this specific occurrence, distinct from the generic ``title``.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI of the specific occurrence (typically the request path); helps correlate with logs.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code repeated in the body so the payload is self-contained for clients that drop the status line.",
            "maximum": 599,
            "minimum": 100,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem class. Relative URIs resolve under the API host; use ``about:blank`` for generic.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "contact@protea.example.org",
      "name": "PROTEA Team"
    },
    "description": "**PROTEA** \u2014 Protein Representation and Ontology-Term Enrichment Analysis.\n\nManages the full pipeline from UniProt sequence ingestion through GPU embedding computation (ESM-2, ESM3c, T5) to KNN-based GO term prediction.\n\nAll long-running operations are queued via RabbitMQ and tracked as `Job` rows with a full event audit trail. Use `GET /jobs/{id}/events` to stream real-time progress.",
    "title": "PROTEA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/embeddings/configs": {
      "get": {
        "description": "List all embedding configurations with their stored embedding counts, newest first.\n\nThe per-config GROUP BY over a 4M-row table is cached 5 minutes; new\nconfigs still appear immediately (they have 0 embeddings), only the\ncounts are stale.",
        "operationId": "list_embedding_configs_v1_embeddings_configs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Embedding Configs V1 Embeddings Configs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List embedding configs",
        "tags": [
          "embeddings"
        ]
      },
      "post": {
        "description": "Create a new EmbeddingConfig that defines the model, layer selection, pooling strategy, and chunking.\n\nThis config is referenced by `compute_embeddings` jobs and `predict_go_terms` jobs to ensure\nquery and reference embeddings were produced under identical settings.",
        "operationId": "create_embedding_config_v1_embeddings_configs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Embedding Config V1 Embeddings Configs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an embedding config",
        "tags": [
          "embeddings"
        ]
      }
    },
    "/v1/embeddings/configs/{config_id}": {
      "delete": {
        "description": "Delete an EmbeddingConfig and cascade-delete all linked embeddings, prediction sets, and predictions.",
        "operationId": "delete_embedding_config_v1_embeddings_configs__config_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Embedding Config V1 Embeddings Configs  Config Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an embedding config",
        "tags": [
          "embeddings"
        ]
      },
      "get": {
        "description": "Retrieve a single EmbeddingConfig with its total stored embedding count.",
        "operationId": "get_embedding_config_v1_embeddings_configs__config_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Embedding Config V1 Embeddings Configs  Config Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get embedding config details",
        "tags": [
          "embeddings"
        ]
      }
    },
    "/v1/embeddings/predict": {
      "post": {
        "description": "Queue a `predict_go_terms` job that runs KNN-based GO term transfer.\n\nThe coordinator partitions query proteins into batches, each dispatched to\n`protea.predictions.batch` workers for KNN search (numpy or FAISS) + GO annotation transfer.\nResults are written to a new `PredictionSet` via `protea.predictions.write` workers.\n\nRequired body fields: `embedding_config_id`, `annotation_set_id`, `ontology_snapshot_id`.\nOptional: `query_set_id` (FASTA upload), `limit_per_entry`, `distance_threshold`,\n`batch_size`, `search_backend`. Feature-engineering flags default to True:\n`compute_alignments`, `compute_taxonomy`, `compute_reranker_features` all run\nunless explicitly set to false. `aspect_separated_knn` defaults to true\n(one KNN index per GO aspect to guarantee BPO/MFO/CCO coverage even when\nunified nearest neighbours carry only one aspect).",
        "operationId": "predict_go_terms_v1_embeddings_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Predict Go Terms V1 Embeddings Predict Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger GO term prediction",
        "tags": [
          "embeddings"
        ]
      }
    },
...
    "/v1/embeddings/prediction-sets/{set_id}": {
      "delete": {
        "description": "Delete a prediction set and all its GOPrediction rows.",
        "operationId": "delete_prediction_set_v1_embeddings_prediction_sets__set_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Prediction Set V1 Embeddings Prediction Sets  Set Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a prediction set",
        "tags": [
          "embeddings"
        ]
      },
      "get": {
        "description": "Retrieve a prediction set with total prediction count and per-protein GO term counts.",
        "operationId": "get_prediction_set_v1_embeddings_prediction_sets__set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Prediction Set V1 Embeddings Prediction Sets  Set Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get prediction set details",
        "tags": [
          "embeddings"
        ]
      }
    },
    "/v1/embeddings/prediction-sets/{set_id}/go-terms": {
      "get": {
        "description": "Return the most frequently predicted GO terms grouped by aspect (F/P/C)\nand the total prediction counts per aspect.",
        "operationId": "get_go_term_distribution_v1_embeddings_prediction_sets__set_id__go_terms_get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Set Id",
              "type": "string"
            }
          },
          {
            "description": "Max number of top-N GO terms to return per aspect (``F``/``P``/``C``).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Max number of top-N GO terms to return per aspect (``F``/``P``/``C``).",
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Go Term Distribution V1 Embeddings Prediction Sets  Set Id  Go Terms Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "GO term distribution in a prediction set",
        "tags": [
          "embeddings"
        ]
      }
    },
    "/v1/embeddings/prediction-sets/{set_id}/predictions-cafa.tsv": {
      "get": {
        "description": "Stream predictions in CAFA format: ``protein_accession\\tgo_id\\tscore``.\n\nScore is computed as ``max(0.0, 1.0 - distance)`` so that closer neighbours\nreceive higher confidence scores in the [0, 1] range expected by the\nCAFA evaluator.  One row per (protein, GO term) pair; duplicate GO terms\nfor the same protein are deduplicated keeping the highest score (lowest distance).\n\nPass ``eval_id`` to restrict output to delta proteins only (NK + LK targets),\nwhich is required for a valid CAFA evaluation.",
        "operationId": "download_predictions_cafa_v1_embeddings_prediction_sets__set_id__predictions_cafa_tsv_get",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Set Id",
              "type": "string"
            }
          },
          {
            "description": "Filter to delta proteins from this EvaluationSet.",
            "in": "query",
            "name": "eval_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to delta proteins from this EvaluationSet.",
              "title": "Eval Id"
            }
          },
          {
            "description": "Filter by GO aspect: F, P, or C",
            "in": "query",
            "name": "aspect",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by GO aspect: F, P, or C",
              "title": "Aspect"
            }
          },
          {
            "description": "Only include predictions with distance \u2264 this value",
            "in": "query",
            "name": "max_distance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only include predictions with distance \u2264 this value",
              "title": "Max Distance"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download predictions in CAFA submission format",
        "tags": [
          "embeddings"
        ]
      }
    },
...
  },
...
}